Move system ix details cards
