+++ explorer/src/components/instruction/system/AssignDetailsCard.tsx
@@ -5,32 +5,36 @@ import {
   SignatureResult,
   SystemInstruction
 } from "@solana/web3.js";
-import { lamportsToSolString } from "utils";
 import { displayAddress } from "utils/tx";
-import { InstructionCard } from "./InstructionCard";
+import { InstructionCard } from "../InstructionCard";
 import Copyable from "components/Copyable";
-import { UnknownDetailsCard } from "./UnknownDetailsCard";
+import { UnknownDetailsCard } from "../UnknownDetailsCard";
 
-export function TransferDetailsCard(props: {
+export function AssignDetailsCard(props: {
   ix: TransactionInstruction;
   index: number;
   result: SignatureResult;
 }) {
   const { ix, index, result } = props;
 
-  let transfer;
+  let params;
   try {
-    transfer = SystemInstruction.decodeTransfer(ix);
+    params = SystemInstruction.decodeAssign(ix);
   } catch (err) {
     console.error(err);
     return <UnknownDetailsCard {...props} />;
   }
 
-  const from = transfer.fromPubkey.toBase58();
-  const to = transfer.toPubkey.toBase58();
-  const [fromMeta, toMeta] = ix.keys;
+  const from = params.fromPubkey.toBase58();
+  const [fromMeta] = ix.keys;
+
   return (
-    <InstructionCard ix={ix} index={index} result={result} title="Transfer">
+    <InstructionCard
+      ix={ix}
+      index={index}
+      result={result}
+      title="Assign Account"
+    >
       <tr>
         <td>Program</td>
         <td className="text-right">
@@ -58,25 +62,12 @@ export function TransferDetailsCard(props: {
       </tr>
 
       <tr>
-        <td>
-          <div className="mr-2 d-md-inline">To Address</div>
-          {!toMeta.isWritable && (
-            <span className="badge badge-soft-dark mr-1">Readonly</span>
-          )}
-          {toMeta.isSigner && (
-            <span className="badge badge-soft-dark mr-1">Signer</span>
-          )}
-        </td>
+        <td>Assigned Owner</td>
         <td className="text-right">
-          <Copyable text={to}>
-            <code>{to}</code>
+          <Copyable text={params.programId.toBase58()}>
+            <code>{displayAddress(params.programId)}</code>
           </Copyable>
         </td>
-      </tr>
-
-      <tr>
-        <td>Transfer Amount (SOL)</td>
-        <td className="text-right">{lamportsToSolString(transfer.lamports)}</td>
       </tr>
     </InstructionCard>
   );
