+++ explorer/src/components/instruction/system/CreateWithSeedDetailsCard.tsx
@@ -7,11 +7,11 @@ import {
 } from "@solana/web3.js";
 import { lamportsToSolString } from "utils";
 import { displayAddress } from "utils/tx";
-import { InstructionCard } from "./InstructionCard";
+import { InstructionCard } from "../InstructionCard";
 import Copyable from "components/Copyable";
-import { UnknownDetailsCard } from "./UnknownDetailsCard";
+import { UnknownDetailsCard } from "../UnknownDetailsCard";
 
-export function CreateDetailsCard(props: {
+export function CreateWithSeedDetailsCard(props: {
   ix: TransactionInstruction;
   index: number;
   result: SignatureResult;
@@ -20,7 +20,7 @@ export function CreateDetailsCard(props: {
 
   let params;
   try {
-    params = SystemInstruction.decodeCreateAccount(ix);
+    params = SystemInstruction.decodeCreateWithSeed(ix);
   } catch (err) {
     console.error(err);
     return <UnknownDetailsCard {...props} />;
@@ -28,6 +28,7 @@ export function CreateDetailsCard(props: {
 
   const from = params.fromPubkey.toBase58();
   const newKey = params.newAccountPubkey.toBase58();
+  const baseKey = params.basePubkey.toBase58();
   const [fromMeta, newMeta] = ix.keys;
 
   return (
@@ -35,7 +36,7 @@ export function CreateDetailsCard(props: {
       ix={ix}
       index={index}
       result={result}
-      title="Create Account"
+      title="Create Account w/ Seed"
     >
       <tr>
         <td>Program</td>
@@ -76,6 +77,41 @@ export function CreateDetailsCard(props: {
         <td className="text-right">
           <Copyable text={newKey}>
             <code>{newKey}</code>
+          </Copyable>
+        </td>
+      </tr>
+
+      <tr>
+        <td>
+          <div className="mr-2 d-md-inline">New Address</div>
+          {!newMeta.isWritable && (
+            <span className="badge badge-soft-dark mr-1">Readonly</span>
+          )}
+          {newMeta.isSigner && (
+            <span className="badge badge-soft-dark mr-1">Signer</span>
+          )}
+        </td>
+        <td className="text-right">
+          <Copyable text={newKey}>
+            <code>{newKey}</code>
+          </Copyable>
+        </td>
+      </tr>
+
+      <tr>
+        <td>Base Address</td>
+        <td className="text-right">
+          <Copyable text={baseKey}>
+            <code>{baseKey}</code>
+          </Copyable>
+        </td>
+      </tr>
+
+      <tr>
+        <td>Seed</td>
+        <td className="text-right">
+          <Copyable text={params.seed}>
+            <code>{params.seed}</code>
           </Copyable>
         </td>
       </tr>
