+++ explorer/src/components/instruction/system/NonceAdvanceDetailsCard.tsx
@@ -6,12 +6,11 @@ import {
   SystemInstruction
 } from "@solana/web3.js";
 import { displayAddress } from "utils/tx";
-import { lamportsToSolString } from "utils";
-import { InstructionCard } from "./InstructionCard";
+import { InstructionCard } from "../InstructionCard";
 import Copyable from "components/Copyable";
-import { UnknownDetailsCard } from "./UnknownDetailsCard";
+import { UnknownDetailsCard } from "../UnknownDetailsCard";
 
-export function NonceWithdrawDetailsCard(props: {
+export function NonceAdvanceDetailsCard(props: {
   ix: TransactionInstruction;
   index: number;
   result: SignatureResult;
@@ -20,24 +19,22 @@ export function NonceWithdrawDetailsCard(props: {
 
   let params;
   try {
-    params = SystemInstruction.decodeNonceWithdraw(ix);
+    params = SystemInstruction.decodeNonceAdvance(ix);
   } catch (err) {
     console.error(err);
     return <UnknownDetailsCard {...props} />;
   }
 
   const nonceKey = params.noncePubkey.toBase58();
-  const toKey = params.toPubkey.toBase58();
   const authorizedKey = params.authorizedPubkey.toBase58();
-  const lamports = params.lamports;
-  const [nonceMeta, toMeta, , , authorizedMeta] = ix.keys;
+  const [nonceMeta, , authorizedMeta] = ix.keys;
 
   return (
     <InstructionCard
       ix={ix}
       index={index}
       result={result}
-      title="Withdraw Nonce"
+      title="Advance Nonce"
     >
       <tr>
         <td>Program</td>
@@ -80,28 +77,6 @@ export function NonceWithdrawDetailsCard(props: {
             <code>{authorizedKey}</code>
           </Copyable>
         </td>
-      </tr>
-
-      <tr>
-        <td>
-          <div className="mr-2 d-md-inline">To Address</div>
-          {!toMeta.isWritable && (
-            <span className="badge badge-soft-dark mr-1">Readonly</span>
-          )}
-          {toMeta.isSigner && (
-            <span className="badge badge-soft-dark mr-1">Signer</span>
-          )}
-        </td>
-        <td className="text-right">
-          <Copyable text={toKey}>
-            <code>{toKey}</code>
-          </Copyable>
-        </td>
-      </tr>
-
-      <tr>
-        <td>Withdraw Amount (SOL)</td>
-        <td className="text-right">{lamportsToSolString(lamports)}</td>
       </tr>
     </InstructionCard>
   );
